% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.r
\name{getcellfreqM}
\alias{getcellfreqM}
\title{Calculate cell frequency per sample from Seurat Object}
\usage{
getcellfreqM(sc, clusterlabel = "minorlabel", samplelabel = "sampleID")
}
\arguments{
\item{sc}{Singlecell data in Seurat format}

\item{clusterlabel}{The column in Seurat object sc used to store cell cluster annotation}

\item{samplelabel}{The column in Seurat object sc used to store sample annotation}
}
\value{
Cell frequency Matrix
}
\description{
This function takes Seurat Object as input, and output cell frequency per sample.
}
