% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherDeconvolution.r
\name{otherDeconvolution}
\alias{otherDeconvolution}
\title{Wrapper function for other deconvolution methods}
\usage{
otherDeconvolution(
  bulk,
  sc.basis,
  markers,
  sc,
  method = c("CIBERSORT", "DSA", "DeconRNASeq", "nnls", "MuSiC", "BisqueRNA"),
  clusterlabel = "minorlabel",
  samplelabel = "sampleID"
)
}
\arguments{
\item{bulk}{Gene by sample TPM matrix, output from prepareInput() function}

\item{sc.basis}{Reference mean and variance gene profile for each cell type, output from prepareInput() function}

\item{markers}{Cell markers for each minor and major cell clusters, output from prepareInput() function}

\item{method}{One of the following deconvolution methods: c("CIBERSORT","DSA","DeconRNASeq","nnls","MuSiC","BisqueRNA")}
}
\value{
Estimated cell frequency matrix
}
\description{
This function takes bulk expression profile, sc reference profile, and marker genes to estimate cell frequency for each sample in the bulk expression.
}
