% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing.r
\name{prepareInput}
\alias{prepareInput}
\title{Prepare input files for ProM}
\usage{
prepareInput(
  bulk,
  sc,
  outfile,
  cellN = 2000,
  majorlabel = "majorlabel",
  minorlabel = "minorlabel",
  samplelabel = "sampleID",
  markerN = 200
)
}
\arguments{
\item{bulk}{Gene by sample TPM/count matrix for deconvolution}

\item{sc}{Singlecell data in Seurat format}

\item{outfile}{Rfile used to store R data output from this function}

\item{cellN}{downsample cellN to speed up calculation of markers}

\item{majorlabel}{The column in Seurat object sc used to store major cluster annotation}

\item{minorlabel}{The column in Seurat object sc used to store minor cluster annotation}

\item{samplelabel}{The column in Seurat object sc used to store sample annotation}

\item{markerN}{The maxium number of markers per cell cluster}
}
\value{
if no outfile supplied, it will return a list of four objects. majorsc.basis: reference mean and variance matrix for major cluster.minorsc.basis: reference mean and variance matrix for minor cluster.bulk:the bulk matrix.markers: cell markers for each major and minor cell clusters.
}
\description{
This function takes the input bulk expression and scRNAseq used for reference. It renormalizes the bulk and sc expression using common genes,generates the mean and variance reference matrix from sc data and computes the marker genes for each cell cluster.
}
