% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProMfunctions.r
\name{ProM}
\alias{ProM}
\title{Wrapper function for ProM deconvolution}
\usage{
ProM(
  bulk,
  sc.basis,
  markers,
  priorcellfreqM,
  method = "combined",
  usebulkSigma = TRUE,
  Zcutoff = 0.5,
  transformation = "none"
)
}
\arguments{
\item{bulk}{Gene by sample TPM matrix, output from prepareInput() function}

\item{sc.basis}{Reference mean and variance gene profile for each cell type, output from prepareInput() function}

\item{markers}{Cell markers for each minor and major cell clusters, output from prepareInput() function}

\item{priorcellfreqM}{Prior cell frequency distribution, can be calculated using getcellfreqM() function}

\item{method}{One of the three ProM stratigies: 1)using both reference profile and cell marker based deconvolution(combined); 2)using only reference profile based deconvolution(profileOnly);3)using only cell marker based deconvolution(markerOnly)}

\item{usebulkSigma}{Indicating whether gene weight/variance that captures the difference between sc and bulk are considered}

\item{Zcutoff}{Cell clusters with Zfrequency (percentage of samples without this cell cluster) larger than Zcutoff need to use linear regression with intercept in marker-based deconvolution.}

\item{transformation}{Whether bulk gene expression need to transform to reduce the possible platform difference between sc and bulk RNAseq. 1)none: no transformation; 2)profile:transformation only applies to reference profile-based deconvolution; 3)marker: transformation only applied to marker-based deconvolution; 4)both: transformation applied to both.}
}
\value{
Estimated cell frequency matrix
}
\description{
This function takes bulk expression profile, sc reference profile, and marker genes to estimate cell frequency for each sample in the bulk expression.
}
